# docs/contracts/chat-console.openapi.yaml
openapi: 3.1.0
info:
  title: Sprint Crew console API
  version: 0.1.0
  description: |
    Status: Implemented. Console sessions are durable in SQLite (M1) and reaped after
    CONSOLE_SESSION_TTL_DAYS. Clarify questions are model-generated by the Interpreter
    on the Work lane (ADR 0013) and degrade to a deterministic fallback when the lane
    is cold.
    Contract for the interactive console
    (docs/contracts/chat-console-api.md, ADR 0012, ADR 0013). Routes are live at
    /v1/console/* alongside the existing /sprint/* and /health API.
    Auth: when CONSOLE_API_TOKEN is set, every /v1/console/* and /sprint/* request
    requires Authorization: Bearer <token>, or ?token= as a fallback on any route
    (needed for EventSource, which cannot set headers). Empty token disables auth.
    GET /health is exempt.
    Timeline: GET /events polls a durable, monotonically sequenced log (M2) and
    GET /stream serves the same events over SSE with Last-Event-ID resume (M3).
    Runs are queued one at a time and cancellable (M5, ADR 0014).
# Either scheme satisfies any route: api/auth.py accepts ?token= everywhere, not just on
# the SSE stream, so a generated client should treat both as valid globally.
security:
  - bearerAuth: []
  - tokenQueryAuth: []
paths:
  /v1/console/sessions:
    post:
      operationId: createConsoleSession
      summary: Create a console session (with initial_prompt → clarifying or ready)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConsoleSessionRequest"
            example:
              mode: code
              initial_prompt: Add a /metrics endpoint with request counters
              repo_url: https://github.com/example/service
              target_language: null
      responses:
        "201":
          description: |
            Session created. With initial_prompt the Interpreter (or fallback) runs
            immediately — status is clarifying or ready, not collecting.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleSession"
              example:
                session_id: cs-7f3a
                mode: code
                status: clarifying
                confirmed: false
                repo_url: https://github.com/example/service
                target_language: null
                messages:
                  - role: user
                    content: Add a /metrics endpoint with request counters
                    timestamp: "2026-07-16T09:00:00+00:00"
                intent: null
                clarify_questions:
                  - question_id: q-scope
                    text: Which part of the repo should change?
                    why_asked: null
                    recommended_suggestion_id: null
                    suggestions:
                      - suggestion_id: s-scope-focused
                        label: Only the files needed for this change
                      - suggestion_id: s-scope-broad
                        label: Related modules too
                        detail: including tests and docs touched by the change
                    allow_custom: true
                clarify_answers: []
                sprint_ref: null
                plan_result: null
                error: null
                workspace_status: cloning
                workspace_root: null
                workspace_error: null
                index_status: pending
                index_error: null
                created_at: "2026-07-16T09:00:00+00:00"
                updated_at: "2026-07-16T09:00:00+00:00"
        "422":
          $ref: "#/components/responses/ValidationError"
    get:
      operationId: listConsoleSessions
      summary: Session history, newest first (M8)
      description: |
        Summaries only — no messages, clarify state or plan results. Fetch one session
        by id for the full object.
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 25, minimum: 1, maximum: 100 }
        - name: offset
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: One page of session summaries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleSessionPage"
              example:
                sessions:
                  - session_id: cs-7f3a
                    mode: code
                    status: completed
                    workspace_status: ready
                    title: Add a /metrics endpoint with request counters
                    repo_url: https://github.com/example/service
                    created_at: "2026-07-16T09:00:00+00:00"
                    updated_at: "2026-07-16T09:41:00+00:00"
                total: 1
                next_offset: null
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Shared CONSOLE_API_TOKEN. Omit the Authorization header when the env is
        empty (auth disabled). GET /health never requires this scheme.
    tokenQueryAuth:
      type: apiKey
      in: query
      name: token
      description: |
        The same shared CONSOLE_API_TOKEN as bearerAuth, passed as a ?token= query
        parameter. Accepted on every route but intended only for GET .../stream, where
        browser EventSource cannot set an Authorization header. A query-string token
        may appear in access logs — acceptable for this single-user LAN deployment.
...
  responses:
...
    ValidationError:
      description: Request body failed schema validation (extra=forbid)
      content:
        application/json:
          schema:
            type: object
  schemas:
    ConsoleMode:
      type: string
      enum: [plan, code]
    ConsoleSessionStatus:
      type: string
      description: |
        "queued" (M5) means started but waiting for the single run slot; see
        queue_position. A session may also go straight from ready to running when
        nothing is ahead of it, so clients must handle both.

        "awaiting_review" (M7) is non-terminal and blocks on the *user*: the run is alive
        and parked on a per-file verdict. Read ConsoleDiffPage.review for what it wants
        and POST .../diff/decisions to release it. It expires (see review.expires_at),
        after which the run fails; Stop also works while parked.
      enum:
        [
          collecting,
          clarifying,
          ready,
          queued,
          running,
          awaiting_review,
          completed,
          failed,
          cancelled,
        ]
    ConsoleMessage:
      type: object
      additionalProperties: false
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
          minLength: 1
        message_id:
          type: string
          minLength: 1
          description: |
            Stable identity for one turn (M9). Every ask_started / answer_delta /
            citation / answer_complete / ask_failed event carries this in `detail`,
            so a client can grow the right bubble when more than one answer exists.
        citations:
          type: array
          items:
            $ref: "#/components/schemas/AnswerCitation"
          description: Only ever populated on an assistant turn produced by an ask (M9).
        timestamp:
          type: string
    AnswerCitation:
      type: object
      additionalProperties: false
      required: [path, source]
      description: |
        Where an answer came from (M9). Derived from the Explainer's tool log and the
        answer's own inline `path:line` references, never asked of the model. Fetch the
        file with GET /v1/console/sessions/{id}/files/{path} to render a link.
      properties:
        path:
          type: string
          minLength: 1
          description: Repository-relative path in the session's checkout.
        start_line:
          type: [integer, "null"]
          minimum: 1
        end_line:
          type: [integer, "null"]
          minimum: 1
        source:
          type: string
          description: |
            Which evidence produced it: "read_file" (the answer opened this file),
            "semantic_search" (it was shown this hit), or "answer" (the prose named a
            path that exists but was not opened). A UI may rank them in that order.
...
    ClarifySuggestion:
      type: object
      additionalProperties: false
      required: [suggestion_id, label]
      properties:
        suggestion_id:
          type: string
          minLength: 1
        label:
          type: string
          minLength: 1
        detail:
          type: [string, "null"]
        rationale:
          type: [string, "null"]
          description: What this choice costs or buys. Null on fallback questions.
    ClarifyQuestion:
      type: object
      additionalProperties: false
      required: [question_id, text, suggestions]
      properties:
        question_id:
          type: string
          minLength: 1
        text:
          type: string
          minLength: 1
        suggestions:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ClarifySuggestion"
        allow_custom:
          type: boolean
          default: true
        why_asked:
          type: [string, "null"]
          description: The ambiguity that prompted this question. Null on fallback questions.
        recommended_suggestion_id:
          type: [string, "null"]
          description: >-
            The answer the backend would pick if the user said "just decide". When non-null
            it always matches one of this question's suggestion_id values. Null on fallback
            questions.
    ClarifyAnswer:
      type: object
      additionalProperties: false
      description: Exactly one of selected_suggestion_id or custom_text is non-null.
      required: [question_id]
      properties:
        question_id:
          type: string
          minLength: 1
        selected_suggestion_id:
          type: [string, "null"]
        custom_text:
          type: [string, "null"]
    IntentSummary:
      type: object
      additionalProperties: false
      description: >-
        What the Interpreter understood, echoed back so the user can correct it before
        confirming. Null when the deterministic fallback produced the questions.
      required: [restated_goal]
      properties:
        restated_goal:
          type: string
          minLength: 1
        assumptions:
          type: array
          items:
            type: string
        unknowns:
          type: array
          items:
            type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
    SprintRunRef:
      type: object
      additionalProperties: false
      description: |
        Progress mapping to the live /sprint/* API: backlog_run_id feeds
        GET /sprint/backlog/{run_id}; sprint_session_ids mirror that run's
        session_ids for GET /sprint/session/{session_id}.
      properties:
        backlog_run_id:
          type: [string, "null"]
        sprint_session_ids:
          type: array
          items:
            type: string
    PlanDepth:
      type: string
      description: |
        How much analysis a plan run does (M10). "quick" is the ScrumMaster only —
        a backlog, tens of seconds on a warm lane. "deep" adds a read-only TechLead
        pass per story for real files_to_touch/steps/acceptance_tests, which is a
        model run each (bounded by the server's backlog story cap).
      enum: [quick, deep]
    PlanPreviewStory:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        rationale:
          type: [string, "null"]
        key:
          type: [string, "null"]
          description: Story key; what `depends_on` and `recommended_first` refer to.
        description:
          type: string
          default: ""
        acceptance_criteria:
          type: string
          default: ""
        priority:
          type: integer
          minimum: 0
          default: 0
        depends_on:
          type: array
          items:
            type: string
        estimated_complexity:
          type: [string, "null"]
          enum: [trivial, simple, complex, null]
        files_to_touch:
          type: array
          items:
            type: string
          description: |
            Empty at "quick" depth — no TechLead ran. Check `ConsolePlanResult.depth`
            before rendering an empty list as "no files affected".
        acceptance_tests:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            type: string
        planning_mode:
          type: [string, "null"]
          description: |
            Which TechLead ladder produced the detail: template, static or tool_loop.
            A "template" plan is not repo analysis — say so rather than presenting it
            as one.
    ConsolePlanResult:
      type: object
      additionalProperties: false
      description: Plan-mode outcome — analysis only, never shipped (ADR 0012).
      required: [summary]
      properties:
        summary:
          type: string
          minLength: 1
        stories:
          type: array
          items:
            $ref: "#/components/schemas/PlanPreviewStory"
        depth:
          allOf:
            - $ref: "#/components/schemas/PlanDepth"
          description: Which depth produced this result — i.e. whether a TechLead ran.
        product_brief:
          oneOf:
            - $ref: "#/components/schemas/ProductBrief"
            - type: "null"
        recommended_first:
          type: [string, "null"]
          description: Story key the ScrumMaster would start with.
        generated_at:
          type: string
    ProductBrief:
      type: object
      additionalProperties: false
      required: [title, summary]
      properties:
        title:
          type: string
          minLength: 1
        summary:
          type: string
          minLength: 1
        goals:
          type: array
          items:
            type: string
        constraints:
          type: array
          items:
            type: string
        out_of_scope:
          type: array
          items:
            type: string
    CreateConsoleSessionRequest:
      type: object
      additionalProperties: false
      required: [mode]
      properties:
        mode:
          $ref: "#/components/schemas/ConsoleMode"
        initial_prompt:
          type: [string, "null"]
        repo_url:
          type: [string, "null"]
        target_language:
          type: [string, "null"]
          description: Phase 3 language-specialized lanes; pass null for now.
...
    ConsoleSession:
      type: object
      additionalProperties: false
      required:
        - session_id
        - mode
        - status
        - confirmed
        - messages
        - clarify_questions
        - clarify_answers
        - created_at
        - updated_at
      properties:
        session_id:
          type: string
          minLength: 1
        mode:
          $ref: "#/components/schemas/ConsoleMode"
        status:
          $ref: "#/components/schemas/ConsoleSessionStatus"
        confirmed:
          type: boolean
          default: false
        repo_url:
          type: [string, "null"]
        target_language:
          type: [string, "null"]
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ConsoleMessage"
        intent:
          oneOf:
            - $ref: "#/components/schemas/IntentSummary"
            - type: "null"
        clarify_questions:
          type: array
          items:
            $ref: "#/components/schemas/ClarifyQuestion"
        clarify_answers:
          type: array
          items:
            $ref: "#/components/schemas/ClarifyAnswer"
        clarify_round:
          type: integer
          minimum: 1
          default: 1
          description: |
            Which round the open questions belong to (M9). A message posted while
            clarifying or ready re-runs the Interpreter, which REPLACES the question
            set and increments this. Question ids are round-scoped (`q-{round}-{n}`),
            so answering an id from a previous round gets 409, not 400 — refetch the
            session and answer the current set. Do not treat `clarify_questions` as
            append-only, and do not assume an answered question stays answered.
        prior_clarifications:
          type: array
          items:
            type: string
          description: |
            Resolved Q&A from superseded rounds, kept as context (M9). Still things
            the user told us, but no longer answers — the questions are gone. Render
            as history, not as a form.
        active_ask_id:
          type: [string, "null"]
          description: The in-flight ask, if any (M9). See `ask_in_flight`.
        ask_in_flight:
          type: boolean
          default: false
          description: |
            True while an answer is being generated (M9). Derived from the server's
            run registry on every read rather than stored, so it is false again after
            a restart that killed the task. Disable the composer while it is true;
            `POST /messages` and `DELETE /sessions/{id}` both 409.
        sprint_ref:
          oneOf:
            - $ref: "#/components/schemas/SprintRunRef"
            - type: "null"
        plan_result:
          oneOf:
            - $ref: "#/components/schemas/ConsolePlanResult"
            - type: "null"
          description: |
            Null until a plan run finishes (M10). Plan mode is asynchronous: `POST
            /start` returns 202 with this still null, and the result arrives with the
            `plan_complete` event. Do not read it off the start response.
        plan_run_id:
          type: [string, "null"]
          description: |
            The plan-mode run in the server's registry (M10). Plan mode has no backlog
            run, so this is what "queued"/"running" refers to for a plan session.
        parent_session_id:
          type: [string, "null"]
          description: |
            Set on a session created by `POST /promote` (M10): the plan session whose
            backlog this code run executes. A sidebar can nest the two.
        error:
          type: [string, "null"]
        queue_position:
          type: [integer, "null"]
          minimum: 1
          description: |
            How many runs must finish before this one starts (M5). Non-null only
            while status is "queued", where it is 1 or more; null once the run is
            admitted or the session ends.
        cancel_requested_at:
          type: [string, "null"]
          description: |
            Set when Stop was accepted but the run has not stopped yet (M5). While
            this is non-null and status is still "running", render a pending state.
        workspace_status:
          allOf:
            - $ref: "#/components/schemas/WorkspaceStatus"
          description: |
            Progress of the session's own checkout (M8). Advances independently of
            `status`: a session is usable while its clone is still arriving, so render
            the two separately rather than folding them into one spinner.
        workspace_root:
          type: [string, "null"]
          description: Absolute server path of the clone; null until it is ready.
        workspace_error:
          type: [string, "null"]
        index_status:
          allOf:
            - $ref: "#/components/schemas/IndexStatus"
          description: |
            Progress of the repository index (M8). "skipped" is not an error — there
            was nothing worth indexing, or indexing is switched off.
        index_error:
          type: [string, "null"]
        contract_version:
          type: integer
          default: 2
          description: |
            Bumped when a breaking change lands, so a client can branch instead of
            flag-daying. 1 = everything up to M9. 2 = plan mode is asynchronous (M10).
        created_at:
          type: string
        updated_at:
          type: string
    WorkspaceStatus:
      type: string
      description: |
        M8. "evicted" means the clone was reclaimed by the workspace LRU: the session
        is still readable, its checkout is gone.
      enum: [pending, cloning, ready, failed, evicted]
    IndexStatus:
      type: string
      enum: [pending, indexing, ready, skipped, failed]
    ConsoleSessionSummary:
      type: object
      additionalProperties: false
      required: [session_id, mode, status, workspace_status, created_at, updated_at]
      properties:
        session_id:
          type: string
          minLength: 1
        mode:
          $ref: "#/components/schemas/ConsoleMode"
        status:
          $ref: "#/components/schemas/ConsoleSessionStatus"
        workspace_status:
          $ref: "#/components/schemas/WorkspaceStatus"
        title:
          type: [string, "null"]
          description: First user message, truncated — the session's name in a history list.
        repo_url:
          type: [string, "null"]
        parent_session_id:
          type: [string, "null"]
          description: Set on a promoted code run (M10) — nest it under its plan.
        created_at:
          type: string
        updated_at:
          type: string
    ConsoleSessionPage:
      type: object
      additionalProperties: false
      required: [sessions, total]
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/ConsoleSessionSummary"
        total:
          type: integer
        next_offset:
          type: [integer, "null"]
          description: Offset of the next page, or null when this is the last one.
...